YARN-10553. split TestDistributedShell
YARN-10553. fix checkstyles
