+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-applications/hadoop-yarn-applications-distributedshell/src/test/java/org/apache/hadoop/yarn/applications/distributedshell/TestDSWithMultipleNodeManager.java
@@ -17,42 +17,48 @@
  */
 
 package org.apache.hadoop.yarn.applications.distributedshell;
+
 import java.io.IOException;
-import java.util.concurrent.ConcurrentMap;
-import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.Arrays;
 import java.util.Collection;
 import java.util.HashSet;
-import java.util.Set;
 import java.util.Iterator;
+import java.util.Set;
+import java.util.concurrent.ConcurrentMap;
+import java.util.concurrent.TimeUnit;
+import java.util.concurrent.atomic.AtomicReference;
 
-import org.apache.hadoop.conf.Configuration;
-import org.apache.hadoop.util.Shell;
-import org.apache.hadoop.yarn.api.records.NodeId;
-import org.apache.hadoop.yarn.api.records.ApplicationId;
-import org.apache.hadoop.yarn.client.api.YarnClient;
-import org.apache.hadoop.yarn.conf.YarnConfiguration;
-import org.apache.hadoop.yarn.server.nodemanager.NodeManager;
-import org.apache.hadoop.yarn.server.resourcemanager.rmapp.RMApp;
-import org.apache.hadoop.yarn.server.resourcemanager.rmapp.attempt.RMAppAttempt;
-import org.apache.hadoop.yarn.server.resourcemanager.RMContext;
-import org.apache.hadoop.yarn.server.resourcemanager.nodelabels.RMNodeLabelsManager;
-import static org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration.PREFIX;
-
-import org.apache.hadoop.yarn.server.resourcemanager.scheduler.ResourceScheduler;
-import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacityScheduler;
-import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration;
-import org.apache.hadoop.yarn.util.resource.DominantResourceCalculator;
 import org.junit.After;
 import org.junit.Assert;
 import org.junit.Before;
+import org.junit.Rule;
 import org.junit.Test;
+import org.junit.rules.TestName;
+import org.junit.rules.Timeout;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-
-import org.apache.hadoop.thirdparty.com.google.common.collect.ImmutableMap;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
+
+import org.apache.hadoop.conf.Configuration;
+import org.apache.hadoop.test.GenericTestUtils;
+import org.apache.hadoop.thirdparty.com.google.common.collect.ImmutableMap;
+import org.apache.hadoop.yarn.api.records.ApplicationId;
+import org.apache.hadoop.yarn.api.records.NodeId;
+import org.apache.hadoop.yarn.client.api.YarnClient;
+import org.apache.hadoop.yarn.conf.YarnConfiguration;
+import org.apache.hadoop.yarn.server.nodemanager.NodeManager;
+import org.apache.hadoop.yarn.server.resourcemanager.RMContext;
+import org.apache.hadoop.yarn.server.resourcemanager.nodelabels.RMNodeLabelsManager;
+import org.apache.hadoop.yarn.server.resourcemanager.rmapp.RMApp;
+import org.apache.hadoop.yarn.server.resourcemanager.rmapp.attempt.RMAppAttempt;
+import org.apache.hadoop.yarn.server.resourcemanager.scheduler.ResourceScheduler;
+import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacityScheduler;
+import org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration;
+import org.apache.hadoop.yarn.server.resourcemanager.scheduler.placement.ResourceUsageMultiNodeLookupPolicy;
+import org.apache.hadoop.yarn.util.resource.DominantResourceCalculator;
+
+import static org.apache.hadoop.yarn.server.resourcemanager.scheduler.capacity.CapacitySchedulerConfiguration.PREFIX;
 
 /**
  * Test for Distributed Shell With Multiple Node Managers.
@@ -64,21 +70,26 @@ public class TestDSWithMultipleNodeManager {
   private static final Logger LOG =
       LoggerFactory.getLogger(TestDSWithMultipleNodeManager.class);
 
-  static final int NUM_NMS = 2;
-  TestDistributedShell distShellTest;
-  private final Boolean multiNodePlacementEnabled;
+  private static final int NUM_NMS = 2;
   private static final String POLICY_CLASS_NAME =
-      "org.apache.hadoop.yarn.server.resourcemanager.scheduler.placement."
-      + "ResourceUsageMultiNodeLookupPolicy";
+      ResourceUsageMultiNodeLookupPolicy.class.getName();
+  private final Boolean multiNodePlacementEnabled;
+  @Rule
+  public TestName name = new TestName();
+  @Rule
+  public Timeout globalTimeout =
+      new Timeout(DistributedShellBaseTest.TEST_TIME_OUT,
+          TimeUnit.MILLISECONDS);
+  private DistributedShellBaseTest distShellTest;
+  private Client dsClient;
 
+  public TestDSWithMultipleNodeManager(Boolean multiNodePlacementEnabled) {
+    this.multiNodePlacementEnabled = multiNodePlacementEnabled;
+  }
 
   @Parameterized.Parameters
   public static Collection<Boolean> getParams() {
     return Arrays.asList(false, true);
-  }
-
-  public TestDSWithMultipleNodeManager(Boolean multiNodePlacementEnabled) {
-    this.multiNodePlacementEnabled = multiNodePlacementEnabled;
   }
 
   private YarnConfiguration getConfiguration(
@@ -105,39 +116,47 @@ public class TestDSWithMultipleNodeManager {
 
   @Before
   public void setup() throws Exception {
-    distShellTest = new TestDistributedShell();
+    distShellTest = new TestDSTimelineV10();
     distShellTest.setupInternal(NUM_NMS,
         getConfiguration(multiNodePlacementEnabled));
   }
 
   @After
   public void tearDown() throws Exception {
-    distShellTest.tearDown();
+    if (dsClient != null) {
+      dsClient.sendStopSignal();
+      dsClient = null;
+    }
+    if (distShellTest != null) {
+      distShellTest.tearDown();
+      distShellTest = null;
+    }
   }
 
   private void initializeNodeLabels() throws IOException {
-    RMContext rmContext = distShellTest.yarnCluster.getResourceManager(0).getRMContext();
-
+    RMContext rmContext = distShellTest.getResourceManager(0).getRMContext();
     // Setup node labels
     RMNodeLabelsManager labelsMgr = rmContext.getNodeLabelManager();
-    Set<String> labels = new HashSet<String>();
+    Set<String> labels = new HashSet<>();
     labels.add("x");
     labelsMgr.addToCluserNodeLabelsWithDefaultExclusivity(labels);
 
     // Setup queue access to node labels
-    distShellTest.conf.set(PREFIX + "root.accessible-node-labels", "x");
-    distShellTest.conf.set(PREFIX + "root.accessible-node-labels.x.capacity",
-        "100");
-    distShellTest.conf.set(PREFIX + "root.default.accessible-node-labels", "x");
-    distShellTest.conf.set(PREFIX
+    distShellTest.setConfiguration(PREFIX + "root.accessible-node-labels", "x");
+    distShellTest.setConfiguration(
+        PREFIX + "root.accessible-node-labels.x.capacity", "100");
+    distShellTest.setConfiguration(
+        PREFIX + "root.default.accessible-node-labels", "x");
+    distShellTest.setConfiguration(PREFIX
         + "root.default.accessible-node-labels.x.capacity", "100");
 
-    rmContext.getScheduler().reinitialize(distShellTest.conf, rmContext);
+    rmContext.getScheduler().reinitialize(distShellTest.getConfiguration(),
+        rmContext);
 
     // Fetch node-ids from yarn cluster
     NodeId[] nodeIds = new NodeId[NUM_NMS];
     for (int i = 0; i < NUM_NMS; i++) {
-      NodeManager mgr = distShellTest.yarnCluster.getNodeManager(i);
+      NodeManager mgr = distShellTest.getNodeManager(i);
       nodeIds[i] = mgr.getNMContext().getNodeId();
     }
 
@@ -145,270 +164,328 @@ public class TestDSWithMultipleNodeManager {
     labelsMgr.addLabelsToNode(ImmutableMap.of(nodeIds[1], labels));
   }
 
-  @Test(timeout=90000)
+  @Test
   public void testDSShellWithNodeLabelExpression() throws Exception {
+    NMContainerMonitor containerMonitorRunner = null;
     initializeNodeLabels();
 
-    // Start NMContainerMonitor
-    NMContainerMonitor mon = new NMContainerMonitor();
-    Thread t = new Thread(mon);
-    t.start();
+    try {
+      // Start NMContainerMonitor
+      containerMonitorRunner = new NMContainerMonitor();
+      containerMonitorRunner.start();
 
-    // Submit a job which will sleep for 60 sec
-    String[] args = {
-        "--jar",
-        TestDistributedShell.APPMASTER_JAR,
-        "--num_containers",
-        "4",
-        "--shell_command",
-        "sleep",
-        "--shell_args",
-        "15",
-        "--master_memory",
-        "512",
-        "--master_vcores",
-        "2",
-        "--container_memory",
-        "128",
-        "--container_vcores",
-        "1",
-        "--node_label_expression",
-        "x"
-    };
+      // Submit a job which will sleep for 60 sec
+      String[] args =
+          DistributedShellBaseTest.createArguments(() -> generateAppName(),
+              "--num_containers",
+              "4",
+              "--shell_command",
+              "sleep",
+              "--shell_args",
+              "15",
+              "--master_memory",
+              "512",
+              "--master_vcores",
+              "2",
+              "--container_memory",
+              "128",
+              "--container_vcores",
+              "1",
+              "--node_label_expression",
+              "x"
+          );
 
-    LOG.info("Initializing DS Client");
-    final Client client =
-        new Client(new Configuration(distShellTest.yarnCluster.getConfig()));
-    boolean initSuccess = client.init(args);
-    Assert.assertTrue(initSuccess);
-    LOG.info("Running DS Client");
-    boolean result = client.run();
-    LOG.info("Client run completed. Result=" + result);
+      LOG.info("Initializing DS Client");
+      dsClient =
+          new Client(
+              new Configuration(distShellTest.getYarnClusterConfiguration()));
+      Assert.assertTrue(dsClient.init(args));
+      LOG.info("Running DS Client");
+      boolean result = dsClient.run();
+      LOG.info("Client run completed. Result={}", result);
 
-    t.interrupt();
+      containerMonitorRunner.stopMonitoring();
 
-    // Check maximum number of containers on each NMs
-    int[] maxRunningContainersOnNMs = mon.getMaxRunningContainersReport();
-    // Check no container allocated on NM[0]
-    Assert.assertEquals(0, maxRunningContainersOnNMs[0]);
-    // Check there're some containers allocated on NM[1]
-    Assert.assertTrue(maxRunningContainersOnNMs[1] > 0);
+      // Check maximum number of containers on each NMs
+      int[] maxRunningContainersOnNMs =
+          containerMonitorRunner.getMaxRunningContainersReport();
+      // Check no container allocated on NM[0]
+      Assert.assertEquals(0, maxRunningContainersOnNMs[0]);
+      // Check there are some containers allocated on NM[1]
+      Assert.assertTrue(maxRunningContainersOnNMs[1] > 0);
+    } finally {
+      if (containerMonitorRunner != null) {
+        containerMonitorRunner.stopMonitoring();
+        containerMonitorRunner.join();
+      }
+    }
   }
 
-  @Test(timeout = 90000)
+  @Test
   public void testDistributedShellWithPlacementConstraint()
       throws Exception {
-    NMContainerMonitor mon = new NMContainerMonitor();
-    Thread t = new Thread(mon);
-    t.start();
+    NMContainerMonitor containerMonitorRunner = null;
+    String[] args =
+        DistributedShellBaseTest.createArguments(() -> generateAppName(),
+            "1",
+            "--shell_command",
+            DistributedShellBaseTest.getSleepCommand(15),
+            "--placement_spec",
+            "zk(1),NOTIN,NODE,zk:spark(1),NOTIN,NODE,zk"
+        );
+    try {
+      containerMonitorRunner = new NMContainerMonitor();
+      containerMonitorRunner.start();
 
-    String[] args = {
-        "--jar",
-        distShellTest.APPMASTER_JAR,
-        "1",
-        "--shell_command",
-        distShellTest.getSleepCommand(15),
-        "--placement_spec",
-        "zk(1),NOTIN,NODE,zk:spark(1),NOTIN,NODE,zk"
-    };
-    LOG.info("Initializing DS Client");
-    final Client client =
-        new Client(new Configuration(distShellTest.yarnCluster.getConfig()));
-    boolean initSuccess = client.init(args);
-    Assert.assertTrue(initSuccess);
-    LOG.info("Running DS Client");
-    boolean result = client.run();
-    LOG.info("Client run completed. Result=" + result);
+      LOG.info("Initializing DS Client with args {}", Arrays.toString(args));
+      dsClient =
+          new Client(
+              new Configuration(distShellTest.getYarnClusterConfiguration()));
+      Assert.assertTrue(dsClient.init(args));
+      LOG.info("Running DS Client");
+      boolean result = dsClient.run();
+      LOG.info("Client run completed. Result={}", result);
 
-    t.interrupt();
+      containerMonitorRunner.stopMonitoring();
 
-    ConcurrentMap<ApplicationId, RMApp> apps = distShellTest.yarnCluster.
-        getResourceManager().getRMContext().getRMApps();
-    RMApp app = apps.values().iterator().next();
-    RMAppAttempt appAttempt = app.getAppAttempts().values().iterator().next();
-    NodeId masterNodeId = appAttempt.getMasterContainer().getNodeId();
-    NodeManager nm1 = distShellTest.yarnCluster.getNodeManager(0);
+      ConcurrentMap<ApplicationId, RMApp> apps =
+          distShellTest.getResourceManager().getRMContext().getRMApps();
+      RMApp app = apps.values().iterator().next();
+      RMAppAttempt appAttempt = app.getAppAttempts().values().iterator().next();
+      NodeId masterNodeId = appAttempt.getMasterContainer().getNodeId();
+      NodeManager nm1 = distShellTest.getNodeManager(0);
 
-    int expectedNM1Count = 1;
-    int expectedNM2Count = 1;
-    if (nm1.getNMContext().getNodeId().equals(masterNodeId)) {
-      expectedNM1Count++;
-    } else {
-      expectedNM2Count++;
+      int[] expectedNMsCount = new int[]{1, 1};
+      if (nm1.getNMContext().getNodeId().equals(masterNodeId)) {
+        expectedNMsCount[0]++;
+      } else {
+        expectedNMsCount[1]++;
+      }
+
+      int[] maxRunningContainersOnNMs =
+          containerMonitorRunner.getMaxRunningContainersReport();
+      Assert.assertEquals(expectedNMsCount[0], maxRunningContainersOnNMs[0]);
+      Assert.assertEquals(expectedNMsCount[1], maxRunningContainersOnNMs[1]);
+    } finally {
+      if (containerMonitorRunner != null) {
+        containerMonitorRunner.stopMonitoring();
+        containerMonitorRunner.join();
+      }
     }
-
-    int[] maxRunningContainersOnNMs = mon.getMaxRunningContainersReport();
-    Assert.assertEquals(expectedNM1Count, maxRunningContainersOnNMs[0]);
-    Assert.assertEquals(expectedNM2Count, maxRunningContainersOnNMs[1]);
   }
 
-  @Test(timeout = 90000)
+  @Test
   public void testDistributedShellWithAllocationTagNamespace()
       throws Exception {
-    NMContainerMonitor mon = new NMContainerMonitor();
-    Thread monitorThread = new Thread(mon);
-    monitorThread.start();
+    NMContainerMonitor containerMonitorRunner = null;
+    Client clientB = null;
+    YarnClient yarnClient = null;
 
-    String[] argsA = {
-        "--jar",
-        distShellTest.APPMASTER_JAR,
-        "--shell_command",
-        distShellTest.getSleepCommand(30),
-        "--placement_spec",
-        "bar(1),notin,node,bar"
-    };
-    final Client clientA =
-        new Client(new Configuration(distShellTest.yarnCluster.getConfig()));
-    clientA.init(argsA);
-    final AtomicBoolean resultA = new AtomicBoolean(false);
-    Thread t = new Thread() {
-      public void run() {
+    String[] argsA =
+        DistributedShellBaseTest.createArguments(() -> generateAppName("01"),
+            "--shell_command",
+            DistributedShellBaseTest.getSleepCommand(30),
+            "--placement_spec",
+            "bar(1),notin,node,bar"
+        );
+    String[] argsB =
+        DistributedShellBaseTest.createArguments(() -> generateAppName("02"),
+            "1",
+            "--shell_command",
+            DistributedShellBaseTest.getListCommand(),
+            "--placement_spec",
+            "foo(3),notin,node,all/bar"
+        );
+
+    try {
+      containerMonitorRunner = new NMContainerMonitor();
+      containerMonitorRunner.start();
+      dsClient =
+          new Client(
+              new Configuration(distShellTest.getYarnClusterConfiguration()));
+      dsClient.init(argsA);
+      Thread dsClientRunner = new Thread(() -> {
         try {
-          resultA.set(clientA.run());
+          dsClient.run();
         } catch (Exception e) {
           throw new RuntimeException(e);
         }
+      });
+      dsClientRunner.start();
+
+      NodeId taskContainerNodeIdA;
+      ConcurrentMap<ApplicationId, RMApp> apps;
+      AtomicReference<RMApp> appARef = new AtomicReference<>(null);
+      AtomicReference<NodeId> masterContainerNodeIdARef =
+          new AtomicReference<>(null);
+      int[] expectedNMCounts = new int[]{0, 0};
+
+      waitForExpectedNMsCount(expectedNMCounts, appARef,
+          masterContainerNodeIdARef);
+
+      NodeId nodeA = distShellTest.getNodeManager(0).getNMContext().
+          getNodeId();
+      NodeId nodeB = distShellTest.getNodeManager(1).getNMContext().
+          getNodeId();
+      Assert.assertEquals(2, (expectedNMCounts[0] + expectedNMCounts[1]));
+      if (expectedNMCounts[0] != expectedNMCounts[1]) {
+        taskContainerNodeIdA = masterContainerNodeIdARef.get();
+      } else {
+        taskContainerNodeIdA =
+            masterContainerNodeIdARef.get().equals(nodeA) ? nodeB : nodeA;
       }
-    };
-    t.start();
 
-    NodeId masterContainerNodeIdA;
-    NodeId taskContainerNodeIdA;
-    ConcurrentMap<ApplicationId, RMApp> apps;
-    RMApp appA;
-
-    int expectedNM1Count = 0;
-    int expectedNM2Count = 0;
-    while (true) {
-      if ((expectedNM1Count + expectedNM2Count) < 2) {
-        expectedNM1Count = distShellTest.yarnCluster.getNodeManager(0).
-            getNMContext().getContainers().size();
-        expectedNM2Count = distShellTest.yarnCluster.getNodeManager(1).
-            getNMContext().getContainers().size();
-        continue;
+      clientB =
+          new Client(
+              new Configuration(distShellTest.getYarnClusterConfiguration()));
+      clientB.init(argsB);
+      Assert.assertTrue(clientB.run());
+      containerMonitorRunner.stopMonitoring();
+      apps = distShellTest.getResourceManager().getRMContext().getRMApps();
+      Iterator<RMApp> it = apps.values().iterator();
+      RMApp appB = it.next();
+      if (appARef.get().equals(appB)) {
+        appB = it.next();
       }
-      apps = distShellTest.yarnCluster.getResourceManager().getRMContext().
-          getRMApps();
-      if (apps.isEmpty()) {
-        Thread.sleep(10);
-        continue;
+      LOG.info("Allocation Tag NameSpace Applications are={} and {}",
+          appARef.get().getApplicationId(), appB.getApplicationId());
+
+      RMAppAttempt appAttemptB =
+          appB.getAppAttempts().values().iterator().next();
+      NodeId masterContainerNodeIdB =
+          appAttemptB.getMasterContainer().getNodeId();
+
+      if (nodeA.equals(masterContainerNodeIdB)) {
+        expectedNMCounts[0]++;
+      } else {
+        expectedNMCounts[1]++;
       }
-      appA = apps.values().iterator().next();
-      if (appA.getAppAttempts().isEmpty()) {
-        Thread.sleep(10);
-        continue;
+      if (nodeA.equals(taskContainerNodeIdA)) {
+        expectedNMCounts[1] += 3;
+      } else {
+        expectedNMCounts[0] += 3;
       }
-      RMAppAttempt appAttemptA = appA.getAppAttempts().values().iterator().
-          next();
-      if (appAttemptA.getMasterContainer() == null) {
-        Thread.sleep(10);
-        continue;
+      int[] maxRunningContainersOnNMs =
+          containerMonitorRunner.getMaxRunningContainersReport();
+      Assert.assertEquals(expectedNMCounts[0], maxRunningContainersOnNMs[0]);
+      Assert.assertEquals(expectedNMCounts[1], maxRunningContainersOnNMs[1]);
+
+      try {
+        yarnClient = YarnClient.createYarnClient();
+        yarnClient.init(
+            new Configuration(distShellTest.getYarnClusterConfiguration()));
+        yarnClient.start();
+        yarnClient.killApplication(appARef.get().getApplicationId());
+      } catch (Exception e) {
+        // Ignore Exception while killing a job
+        LOG.warn("Exception killing the job: {}", e.getMessage());
       }
-      masterContainerNodeIdA = appAttemptA.getMasterContainer().getNodeId();
-      break;
-    }
-
-    NodeId nodeA = distShellTest.yarnCluster.getNodeManager(0).getNMContext().
-        getNodeId();
-    NodeId nodeB = distShellTest.yarnCluster.getNodeManager(1).getNMContext().
-        getNodeId();
-    Assert.assertEquals(2, (expectedNM1Count + expectedNM2Count));
-
-    if (expectedNM1Count != expectedNM2Count) {
-      taskContainerNodeIdA = masterContainerNodeIdA;
-    } else {
-      taskContainerNodeIdA = masterContainerNodeIdA.equals(nodeA) ? nodeB :
-          nodeA;
-    }
-
-    String[] argsB = {
-        "--jar",
-        distShellTest.APPMASTER_JAR,
-        "1",
-        "--shell_command",
-        Shell.WINDOWS ? "dir" : "ls",
-        "--placement_spec",
-        "foo(3),notin,node,all/bar"
-    };
-    final Client clientB = new Client(new Configuration(distShellTest.
-        yarnCluster.getConfig()));
-    clientB.init(argsB);
-    boolean resultB = clientB.run();
-    Assert.assertTrue(resultB);
-
-    monitorThread.interrupt();
-    apps = distShellTest.yarnCluster.getResourceManager().getRMContext().
-        getRMApps();
-    Iterator<RMApp> it = apps.values().iterator();
-    RMApp appB = it.next();
-    if (appA.equals(appB)) {
-      appB = it.next();
-    }
-    LOG.info("Allocation Tag NameSpace Applications are=" + appA.
-        getApplicationId() + " and " + appB.getApplicationId());
-
-    RMAppAttempt appAttemptB = appB.getAppAttempts().values().iterator().
-        next();
-    NodeId masterContainerNodeIdB = appAttemptB.getMasterContainer().
-        getNodeId();
-
-    if (nodeA.equals(masterContainerNodeIdB)) {
-      expectedNM1Count += 1;
-    } else {
-      expectedNM2Count += 1;
-    }
-    if (nodeA.equals(taskContainerNodeIdA)) {
-      expectedNM2Count += 3;
-    } else {
-      expectedNM1Count += 3;
-    }
-    int[] maxRunningContainersOnNMs = mon.getMaxRunningContainersReport();
-    Assert.assertEquals(expectedNM1Count, maxRunningContainersOnNMs[0]);
-    Assert.assertEquals(expectedNM2Count, maxRunningContainersOnNMs[1]);
-
-    try {
-      YarnClient yarnClient = YarnClient.createYarnClient();
-      yarnClient.init(new Configuration(distShellTest.yarnCluster.
-          getConfig()));
-      yarnClient.start();
-      yarnClient.killApplication(appA.getApplicationId());
-    } catch (Exception e) {
-     // Ignore Exception while killing a job
+    } finally {
+      if (yarnClient != null) {
+        yarnClient.stop();
+      }
+      if (clientB != null) {
+        clientB.sendStopSignal();
+      }
+      if (containerMonitorRunner != null) {
+        containerMonitorRunner.stopMonitoring();
+        containerMonitorRunner.join();
+      }
     }
   }
 
+  protected String generateAppName() {
+    return generateAppName(null);
+  }
+
+  protected String generateAppName(String postFix) {
+    return name.getMethodName().replaceFirst("test", "")
+        .concat(postFix == null ? "" : "-" + postFix);
+  }
+
+  private void waitForExpectedNMsCount(int[] expectedNMCounts,
+      AtomicReference<RMApp> appARef,
+      AtomicReference<NodeId> masterContainerNodeIdARef) throws Exception {
+    GenericTestUtils.waitFor(() -> {
+      if ((expectedNMCounts[0] + expectedNMCounts[1]) < 2) {
+        expectedNMCounts[0] =
+            distShellTest.getNodeManager(0).getNMContext()
+                .getContainers().size();
+        expectedNMCounts[1] =
+            distShellTest.getNodeManager(1).getNMContext()
+                .getContainers().size();
+        return false;
+      }
+      ConcurrentMap<ApplicationId, RMApp> appIDsMap =
+          distShellTest.getResourceManager().getRMContext().getRMApps();
+      if (appIDsMap.isEmpty()) {
+        return false;
+      }
+      appARef.set(appIDsMap.values().iterator().next());
+      if (appARef.get().getAppAttempts().isEmpty()) {
+        return false;
+      }
+      RMAppAttempt appAttemptA =
+          appARef.get().getAppAttempts().values().iterator().next();
+      if (appAttemptA.getMasterContainer() == null) {
+        return false;
+      }
+      masterContainerNodeIdARef.set(
+          appAttemptA.getMasterContainer().getNodeId());
+      return true;
+    }, 10, 60000);
+  }
+
   /**
-   * Monitor containers running on NMs
+   * Monitor containers running on NMs.
    */
-  class NMContainerMonitor implements Runnable {
+  class NMContainerMonitor extends Thread {
     // The interval of milliseconds of sampling (500ms)
-    final static int SAMPLING_INTERVAL_MS = 500;
+    private final static int SAMPLING_INTERVAL_MS = 500;
 
     // The maximum number of containers running on each NMs
-    int[] maxRunningContainersOnNMs = new int[NUM_NMS];
+    private final int[] maxRunningContainersOnNMs = new int[NUM_NMS];
+    private final Object quitSignal = new Object();
+    private volatile boolean isRunning = true;
 
     @Override
     public void run() {
-      while (true) {
+      while (isRunning) {
         for (int i = 0; i < NUM_NMS; i++) {
           int nContainers =
-              distShellTest.yarnCluster.getNodeManager(i).getNMContext()
+              distShellTest.getNodeManager(i).getNMContext()
                   .getContainers().size();
           if (nContainers > maxRunningContainersOnNMs[i]) {
             maxRunningContainersOnNMs[i] = nContainers;
           }
         }
-        try {
-          Thread.sleep(SAMPLING_INTERVAL_MS);
-        } catch (InterruptedException e) {
-          e.printStackTrace();
-          break;
+        synchronized (quitSignal) {
+          try {
+            if (!isRunning) {
+              break;
+            }
+            quitSignal.wait(SAMPLING_INTERVAL_MS);
+          } catch (InterruptedException e) {
+            LOG.warn("NMContainerMonitor interrupted");
+            isRunning = false;
+            break;
+          }
         }
       }
     }
 
     public int[] getMaxRunningContainersReport() {
       return maxRunningContainersOnNMs;
+    }
+
+    public void stopMonitoring() {
+      if (!isRunning) {
+        return;
+      }
+      synchronized (quitSignal) {
+        isRunning = false;
+        quitSignal.notifyAll();
+      }
     }
   }
 }
